submit 관련 매핑함수를 다른 클래스로 분리
